requirements.txt
alembic==1.7.5
backcall==0.2.0
click==8.0.3
decorator==5.1.0
environs==9.3.5
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
SQLAlchemy==1.4.27
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
